# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RentaCar Operational Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      parameters:
        - name: status_reserva
          in: query
          schema:
            type: string
        - name: status_pagamento
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by pickup/return date (YYYY-MM-DD)
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a new pre-reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created reservation with payment intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationCreated"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          description: Not found
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update reservation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}/simulate-payment:
    post:
      operationId: simulatePayment
      tags: [reservations]
      summary: Simulate a payment confirmation for demo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment simulated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}/upload-docs:
    post:
      operationId: uploadDocs
      tags: [uploads]
      summary: Upload check-in documents for a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocsUpload"
      responses:
        "200":
          description: Documents uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}/upload-car-photos:
    post:
      operationId: uploadCarPhotos
      tags: [uploads]
      summary: Upload car condition photos for a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarPhotosUpload"
      responses:
        "200":
          description: Photos uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /sos:
    get:
      operationId: listSosAlerts
      tags: [sos]
      summary: List SOS alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ativo, resolvido]
      responses:
        "200":
          description: List of SOS alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SosAlert"
    post:
      operationId: createSosAlert
      tags: [sos]
      summary: Create a new SOS alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SosAlertInput"
      responses:
        "201":
          description: SOS alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosAlert"
  /sos/{id}/resolve:
    patch:
      operationId: resolveSosAlert
      tags: [sos]
      summary: Mark a SOS alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SOS alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosAlert"
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, marca_modelo, categoria, preco_base_dia, valor_caucao, extra_franquia_zero, status]
      properties:
        id:
          type: integer
        marca_modelo:
          type: string
        categoria:
          type: string
          enum: [Economico, Familiar, SUV, Premium]
        foto_url:
          type: ["string", "null"]
        preco_base_dia:
          type: number
        valor_caucao:
          type: number
        extra_franquia_zero:
          type: number
        status:
          type: string
          enum: [disponivel, alugado, manutencao, reservado_temporario]
        created_at:
          type: ["string", "null"]
...
    Reservation:
      type: object
      required: [id, cliente_telefone, cliente_nome, veiculo_id, tipo_protecao, data_levantamento, data_devolucao, valor_total, status_pagamento, status_reserva]
      properties:
        id:
          type: integer
        cliente_telefone:
          type: string
        cliente_nome:
          type: string
        cliente_idioma:
          type: ["string", "null"]
        veiculo_id:
          type: integer
        tipo_protecao:
          type: string
          enum: [standard_com_caucao, franquia_zero]
        data_levantamento:
          type: string
        data_devolucao:
          type: string
        hora_chegada_voo:
          type: ["string", "null"]
        taxa_noturna_aplicada:
          type: ["number", "null"]
        valor_total:
          type: number
        status_pagamento:
          type: string
          enum: [pendente, pago_sinal, falhado]
        status_reserva:
          type: string
          enum: [criada, checkin_feito, carro_na_estrada, concluida]
        docs_checkin_url:
          type: ["object", "null"]
        fotos_estado_carro:
          type: ["object", "null"]
        stripe_intent_id:
          type: ["string", "null"]
        created_at:
          type: ["string", "null"]
        vehicle:
          $ref: "#/components/schemas/Vehicle"
    ReservationInput:
      type: object
      required: [cliente_telefone, cliente_nome, veiculo_id, tipo_protecao, data_levantamento, data_devolucao]
      properties:
        cliente_telefone:
          type: string
        cliente_nome:
          type: string
        cliente_idioma:
          type: string
        veiculo_id:
          type: integer
        tipo_protecao:
          type: string
          enum: [standard_com_caucao, franquia_zero]
        data_levantamento:
          type: string
        data_devolucao:
          type: string
        hora_chegada_voo:
          type: string
    ReservationCreated:
      type: object
      required: [reservation, payment_url]
      properties:
        reservation:
          $ref: "#/components/schemas/Reservation"
        payment_url:
          type: string
    ReservationUpdate:
      type: object
      properties:
        status_reserva:
          type: string
          enum: [criada, checkin_feito, carro_na_estrada, concluida]
        status_pagamento:
          type: string
          enum: [pendente, pago_sinal, falhado]
        hora_chegada_voo:
          type: string
    DocsUpload:
      type: object
      required: [docs]
      properties:
        docs:
          type: object
          description: JSON object with document URLs (passaporte, carta_conducao)
    CarPhotosUpload:
      type: object
      required: [fotos]
      properties:
        fotos:
          type: object
          description: JSON object with car photo URLs (frente, traseira, esquerda, direita)
    SosAlert:
      type: object
      required: [id, reserva_id, status, created_at]
      properties:
        id:
          type: integer
        reserva_id:
          type: integer
        localizacao_latitude:
          type: ["number", "null"]
        localizacao_longitude:
          type: ["number", "null"]
        foto_problema_url:
          type: ["string", "null"]
        status:
          type: string
          enum: [ativo, resolvido]
        created_at:
          type: string
        reservation:
          $ref: "#/components/schemas/Reservation"
    SosAlertInput:
      type: object
      required: [reserva_id]
      properties:
        reserva_id:
          type: integer
        localizacao_latitude:
          type: number
        localizacao_longitude:
          type: number
        foto_problema_url:
          type: string
...