# lib/api-zod/src/generated/types/docsUpload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentaCar Operational Management API
 * OpenAPI spec version: 0.1.0
 */
import type { DocsUploadDocs } from './docsUploadDocs';

export interface DocsUpload {
  /** JSON object with document URLs (passaporte, carta_conducao) */
  docs: DocsUploadDocs;
}