# lib/api-zod/src/generated/types/sosAlertInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentaCar Operational Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SosAlertInput {
  reserva_id: number;
  localizacao_latitude?: number;
  localizacao_longitude?: number;
  foto_problema_url?: string;
}